pyproject.toml
[project]
name = "fabricmcp_server"
version = "0.1.0"
description = "MCP Server for Microsoft Fabric"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]
dependencies = [
    "mcp[cli]",
    "fastmcp",
    "fastapi[standard]",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "azure-identity",
    "cachetools",
    "aiohttp"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "uv>=0.1.40",
]

[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fabricmcp_server"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "A"]